scripts/requirements.txt
# Python dependencies for Stripe scripts
stripe>=8.0.0
python-dotenv>=1.0.0
